requirements.txt
requests
beautifulsoup4
pandas
cloudscraper
boto3
PyExecJS

